requirements.txt
yfinance==0.2.49
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
pydantic==2.10.3
pydantic-settings==2.6.1
PyYAML==6.0.2
